+++ main.go
@@ -11,6 +11,8 @@ import (
 	"os"
 	"os/signal"
 	"path/filepath"
+	"regexp"
+	"strings"
 	"syscall"
 	"time"
 
@@ -54,11 +56,12 @@ type RegistryCmd struct {
 	DataDir                      string        `arg:"--data-dir,env:DATA_DIR" default:"/var/lib/spegel" help:"Directory where Spegel persists data."`
 	RouterAddr                   string        `arg:"--router-addr,env:ROUTER_ADDR" default:":5001" help:"address to serve router."`
 	RegistryAddr                 string        `arg:"--registry-addr,env:REGISTRY_ADDR" default:":5000" help:"address to server image registry."`
-	MirroredRegistries           []string      `arg:"--mirrored-registries,env:MIRRORED_REGISTRIES" help:"Registries that are configured to be mirrored, if slice is empty all registires are mirrored."`
+	MirroredRegistries           []string      `arg:"--mirrored-registries,env:MIRRORED_REGISTRIES" help:"Registries that are configured to be mirrored, if slice is empty all registries are mirrored."`
+	RegistryFilters              []string      `arg:"--registry-filters,env:REGISTRY_FILTERS" help:"Regular expressions to filter out tags/registries, if slice is empty all registries/tags are resolved."`
 	MirrorResolveTimeout         time.Duration `arg:"--mirror-resolve-timeout,env:MIRROR_RESOLVE_TIMEOUT" default:"20ms" help:"Max duration spent finding a mirror."`
 	MirrorResolveRetries         int           `arg:"--mirror-resolve-retries,env:MIRROR_RESOLVE_RETRIES" default:"3" help:"Max amount of mirrors to attempt."`
-	ResolveLatestTag             bool          `arg:"--resolve-latest-tag,env:RESOLVE_LATEST_TAG" default:"true" help:"When true latest tags will be resolved to digests."`
 	DebugWebEnabled              bool          `arg:"--debug-web-enabled,env:DEBUG_WEB_ENABLED" default:"false" help:"When true enables debug web page."`
+	ResolveLatestTag             bool          `arg:"--resolve-latest-tag,env:RESOLVE_LATEST_TAG" default:"true" help:"When true latest tags will be resolved to digests."`
 }
 
 type CleanupCmd struct {
@@ -164,22 +167,27 @@ func registryCommand(ctx context.Context, args *RegistryCmd) (err error) {
 	if err != nil {
 		return err
 	}
+
+	filters, err := validateFilters(args.RegistryFilters)
+	if err != nil {
+		return err
+	}
 	g.Go(func() error {
 		return router.Run(ctx)
 	})
 
 	// State tracking
 	g.Go(func() error {
-		err := state.Track(ctx, ociStore, router, args.ResolveLatestTag)
+		err := state.Track(ctx, ociStore, router, filters, args.ResolveLatestTag)
 		if err != nil && !errors.Is(err, context.Canceled) {
 			return err
 		}
 		return nil
 	})
-
 	// Registry
 	registryOpts := []registry.RegistryOption{
 		registry.WithResolveLatestTag(args.ResolveLatestTag),
+		registry.WithRegistryFilters(filters),
 		registry.WithResolveRetries(args.MirrorResolveRetries),
 		registry.WithResolveTimeout(args.MirrorResolveTimeout),
 		registry.WithBasicAuth(username, password),
@@ -291,4 +299,23 @@ func loadBasicAuth() (string, string, error) {
 		return "", "", err
 	}
 	return string(username), string(password), nil
+}
+
+func validateFilters(patterns []string) ([]*regexp.Regexp, error) {
+	var filters []*regexp.Regexp
+	// Try to match against each pattern
+	for _, pattern := range patterns {
+		// Skip empty patterns
+		if strings.TrimSpace(pattern) == "" {
+			continue
+		}
+		// Compile pattern and match
+		compiled, err := regexp.Compile(pattern)
+		if err != nil {
+			return nil, err
+		}
+		filters = append(filters, compiled)
+	}
+
+	return filters, nil
 }